refactor: 모임 가입 부분 및 카드리스트 타입 분리 #64
