+++ src/components/RegularClubList.tsx
@@ -6,29 +6,7 @@ import Link from "next/link";
 import { getRegularClubList } from "@/app/(pages)/(club)/club/_api/supabase";
 import Image from "next/image";
 import { useAuth } from "@/app/store/AuthContext";
-
-interface RegularClubForm {
-  m_c_id: number;
-  r_c_member: { count: number }[];
-  regular_club_age: number;
-  regular_club_approbval: boolean;
-  regular_club_create_at: string;
-  regular_club_id: number;
-  regular_club_image: string;
-  regular_club_introduction: string;
-  regular_clun_name: string;
-  regular_club_people_limited: number;
-  s_c_id: number;
-  user_id: {
-    user_name: string;
-    user_profile_img: string;
-  };
-  wish_list: {
-    r_c_id: number;
-    user_id: string;
-    wish_list_id: number;
-  }[];
-}
+import { EggClubForm } from "@/types/cardlist.types";
 
 const RegularClubList = () => {
   const {
@@ -44,7 +22,7 @@ const RegularClubList = () => {
   // console.log(userId);
   // console.log(list);
 
-  const renderHeartIcon = (club: RegularClubForm) => {
+  const renderHeartIcon = (club: EggClubForm) => {
     // console.log("club", club);
 
     if (!userId) {
